feat: 중복 예약 검증 로직 추가

+++ src/main/java/roomescape/reservationTime/service/ReservationTimeService.java
@@ -3,6 +3,7 @@ package roomescape.reservationTime.service;
 import java.util.List;
 import org.springframework.stereotype.Service;
 import roomescape.common.Dao;
+import roomescape.reservation.domain.Reservation;
 import roomescape.reservationTime.domain.ReservationTime;
 import roomescape.reservationTime.dto.ReservationTimeRequest;
 import roomescape.reservationTime.dto.ReservationTimeResponse;
@@ -10,9 +11,11 @@ import roomescape.reservationTime.dto.ReservationTimeResponse;
 @Service
 public class ReservationTimeService {
     private final Dao<ReservationTime> reservationTimeDao;
+    private final Dao<Reservation> reservationDao;
 
-    public ReservationTimeService(Dao<ReservationTime> reservationTimeDao) {
+    public ReservationTimeService(Dao<ReservationTime> reservationTimeDao, Dao<Reservation> reservationDao) {
         this.reservationTimeDao = reservationTimeDao;
+        this.reservationDao = reservationDao;
     }
 
     public ReservationTimeResponse add(ReservationTimeRequest reservationTimeRequest) {
@@ -28,7 +31,16 @@ public class ReservationTimeService {
                 .toList();
     }
 
+    // TODO 고민해보기 dao vs service
     public void deleteById(Long id) {
+        List<Reservation> reservations = reservationDao.findAll();
+
+        boolean isOccupiedId = reservations.stream()
+                        .anyMatch(reservation -> reservation.getTime().getId().equals(id));
+        if (isOccupiedId) {
+            throw new IllegalArgumentException("[ERROR] 이미 예약된 시간은 삭제할 수 없습니다");
+        }
+
         reservationTimeDao.deleteById(id);
     }
 }
